+++ android/src/main/java/com/ledger/reactnative/RCTCoreOperation.java
@@ -377,7 +377,7 @@ public class RCTCoreOperation extends ReactContextBaseJavaModule {
     }
     /**
      * Get block height on which operation was included.
-     * @return Optional 64-bit integer, height of block in which operation was validated
+     * @return Optional 32-bit integer, height of block in which operation was validated
      */
     @ReactMethod
     public void getBlockHeight(ReadableMap currentInstance, Promise promise) {
@@ -387,14 +387,14 @@ public class RCTCoreOperation extends ReactContextBaseJavaModule {
 
             Operation currentInstanceObj = this.javaObjects.get(sUid);
 
-            Long javaResult = currentInstanceObj.getBlockHeight();
+            Integer javaResult = currentInstanceObj.getBlockHeight();
             WritableNativeMap result = new WritableNativeMap();
             if (javaResult == null)
             {
                 promise.resolve(javaResult);
                 return;
             }
-            result.putDouble("value", javaResult);
+            result.putInt("value", javaResult);
 
             promise.resolve(result);
         }
