+++ ios/Sources/include/Operation.hpp
@@ -92,9 +92,9 @@ public:
 
     /**
      * Get block height on which operation was included.
-     * @return Optional 64-bit integer, height of block in which operation was validated
+     * @return Optional 32-bit integer, height of block in which operation was validated
      */
-    virtual std::experimental::optional<int64_t> getBlockHeight() = 0;
+    virtual std::experimental::optional<int32_t> getBlockHeight() = 0;
 
     /**
      * Convert operation as Bitcoin operation.
